add ability to edit early created templates

+++ plugins/templates-generator/admin/src/containers/App/saga.js
@@ -16,10 +16,12 @@ import {
     userLoadingError,
     genresLoadingSuccess,
     genresLoadingError,
-    uploadSuccess,
-    uploadError,
+    previewSuccess,
+    previewError,
     saveError,
     saveSuccess,
+    updateSuccess,
+    updateError,
     getTemplatesError,
     getTemplatesSuccess,
     removeTemplateError,
@@ -31,8 +33,9 @@ import constants from './constants';
 const {
     USER_LOADING_REQUEST,
     GENRES_LOADING_REQUEST,
-    UPLOAD_REQUEST,
+    PREVIEW_REQUEST,
     SAVE_REQUEST,
+    UPDATE_REQUEST,
     TEMPLATES_REQUEST,
     TEMPLATES_REMOVE_REQUEST,
 } = constants;
@@ -67,7 +70,7 @@ export function* getGenres() {
     }
 }
 
-function* uploadFile(action) {
+function* getPreview(action) {
     try {
         const headers = {
             'X-Forwarded-Host': 'strapi',
@@ -85,15 +88,15 @@ function* uploadFile(action) {
         );
 
         if (response) {
-            yield put(uploadSuccess(response));
+            yield put(previewSuccess(response));
         } else {
-            yield put(uploadError());
+            yield put(previewError());
         }
     } catch (err) {
         if (window.strapi) {
             window.strapi.notification.error('notification.error');
         }
-        yield put(uploadError(err));
+        yield put(previewError(err));
     }
 }
 
@@ -102,28 +105,132 @@ function* save(action) {
         const headers = {
             'X-Forwarded-Host': 'strapi',
         };
-        const formData = new FormData();
-        Object.keys(action.payload).forEach((item) => {
-            formData.append(item, action.payload[item]);
-        });
-        const response = yield call(
-            request,
-            '/templates-generator',
-            { method: 'POST', headers, body: formData },
-            false,
-            false
-        );
 
-        if (response) {
-            yield put(saveSuccess(response));
+        const originalFileData = new FormData();
+        originalFileData.append('files', action.payload.originalFile);
+        const templatePreviewData = new FormData();
+        templatePreviewData.append('files', action.payload.templatePreview);
+
+        const [originalFileResponse, templatePreviewResponse] = yield all([
+            call(
+                request,
+                '/upload',
+                { method: 'POST', headers, body: originalFileData },
+                false,
+                false
+            ),
+            call(
+                request,
+                '/upload',
+                { method: 'POST', headers, body: templatePreviewData },
+                false,
+                false
+            ),
+        ]);
+
+        if (
+            originalFileResponse
+            && originalFileResponse[0]
+            && templatePreviewResponse
+            && templatePreviewResponse[0]
+        ) {
+            const originalFile = originalFileResponse[0];
+            const templatePreview = templatePreviewResponse[0];
+            const templateData = new FormData();
+            Object.keys(action.payload.template).forEach((item) => {
+                templateData.append(item, action.payload.template[item]);
+            });
+            templateData.append(
+                'originalImage',
+                `/uploads/${originalFile.hash}${originalFile.ext}`
+            );
+            templateData.append('originalFileId', originalFile.id);
+            templateData.append('image', `/uploads/${templatePreview.hash}${templatePreview.ext}`);
+            templateData.append('fileId', templatePreview.id);
+            templateData.append('system', 'true');
+            const templateResponse = yield call(
+                request,
+                '/content-manager/explorer/templates',
+                { method: 'POST', headers, body: templateData },
+                false,
+                false
+            );
+
+            if (templateResponse) {
+                yield put(saveSuccess(templateResponse));
+            } else {
+                yield put(saveError('can\'t save template'));
+            }
         } else {
-            yield put(saveError());
+            yield put(saveError('can\'t save file'));
         }
     } catch (err) {
         if (window.strapi) {
             window.strapi.notification.error('notification.error');
         }
         yield put(saveError(err));
+    }
+}
+
+function* update(action) {
+    try {
+        const headers = {
+            'X-Forwarded-Host': 'strapi',
+        };
+
+        const templatePreviewData = new FormData();
+        templatePreviewData.append('files', action.payload.templatePreview);
+
+        const [prevPreviewResponse, templatePreviewResponse] = yield all([
+            call(
+                request,
+                `/upload/files/${action.payload.template.fileId}`,
+                { method: 'DELETE' }
+            ),
+            call(
+                request,
+                '/upload',
+                { method: 'POST', headers, body: templatePreviewData },
+                false,
+                false
+            ),
+        ]);
+
+        if (
+            prevPreviewResponse
+            && templatePreviewResponse
+            && templatePreviewResponse[0]
+        ) {
+            const templatePreview = templatePreviewResponse[0];
+            const templateData = new FormData();
+            Object.keys(action.payload.template).forEach((item) => {
+                if (item !== 'fileId' && item !== 'image') {
+                    templateData.append(item, action.payload.template[item]);
+                }
+            });
+            templateData.append('image', `/uploads/${templatePreview.hash}${templatePreview.ext}`);
+            templateData.append('fileId', templatePreview.id);
+            const templateResponse = yield call(
+                request,
+                `/content-manager/explorer/templates/${action.payload.template.id}`,
+                { method: 'PUT', headers, body: templateData },
+                false,
+                false
+            );
+
+            if (templateResponse) {
+                yield put(updateSuccess(templateResponse));
+            } else {
+                yield put(updateError('can\'t save template'));
+            }
+        } else {
+            yield put(updateError('can\'t save file'));
+        }
+    } catch (err) {
+        if (window.strapi) {
+            window.strapi.notification.error('notification.error');
+        }
+        yield put(updateError(err));
     }
 }
 
@@ -145,14 +252,16 @@ export function* getTemplates(genre) {
 export function* removeTemplate(action) {
     try {
         const fileURL = `/upload/files/${action.payload.fileId}`;
+        const originalFileURL = `/upload/files/${action.payload.originalFileId}`;
         const templateURL = `/content-manager/explorer/templates/${action.payload.templateId}`;
 
-        const [file, template] = yield all([
+        const [file, originalFile, template] = yield all([
             call(request, fileURL, { method: 'DELETE' }),
+            call(request, originalFileURL, { method: 'DELETE' }),
             call(request, templateURL, { method: 'DELETE' }),
         ]);
 
-        if (file && template) {
+        if (file && originalFile && template) {
             yield put(removeTemplateSuccess(template));
             strapi.notification.success('templates-generator.Preview.Remove-Success');
         } else {
@@ -166,8 +275,9 @@ export function* removeTemplate(action) {
 export default function* defaultSaga() {
     const getUser = yield fork(takeLatest, USER_LOADING_REQUEST, userGet);
     const getGenresRequest = yield fork(takeLatest, GENRES_LOADING_REQUEST, getGenres);
-    const filUploadRequest = yield fork(takeLatest, UPLOAD_REQUEST, uploadFile);
+    const getPreviewRequest = yield fork(takeLatest, PREVIEW_REQUEST, getPreview);
     const saveRequest = yield fork(takeLatest, SAVE_REQUEST, save);
+    const updateRequest = yield fork(takeLatest, UPDATE_REQUEST, update);
     const getTemplatesRequest = yield fork(takeLatest, TEMPLATES_REQUEST, getTemplates);
     const removeTemplateRequest = yield fork(takeLatest, TEMPLATES_REMOVE_REQUEST, removeTemplate);
 
@@ -175,8 +285,9 @@ export default function* defaultSaga() {
     yield take(LOCATION_CHANGE);
     yield cancel(getUser);
     yield cancel(getGenresRequest);
-    yield cancel(filUploadRequest);
+    yield cancel(getPreviewRequest);
     yield cancel(saveRequest);
+    yield cancel(updateRequest);
     yield cancel(getTemplatesRequest);
     yield cancel(removeTemplateRequest);
 }
